test: 修复 Long 与 Int 类型混用导致断言失败，全部改为 Long 字面量

+++ app/src/test/kotlin/com/lanxin/android/builtin/knowledge/BertTokenizerTest.kt
@@ -14,57 +14,39 @@ class BertTokenizerTest {
         val encoding = tokenizer.encode("hello world")
         // [CLS]=101, hello=7592, world=2088, [SEP]=102, then [PAD]=0 padded
         assertEquals(512, encoding.inputIds.size)
-        assertEquals(101, encoding.inputIds[0])
-        assertEquals(7592, encoding.inputIds[1])
-        assertEquals(2088, encoding.inputIds[2])
-        assertEquals(102, encoding.inputIds[3])
-        assertEquals(0, encoding.inputIds[4])  // [PAD]
+        assertEquals(101L, encoding.inputIds[0])
+        assertEquals(7592L, encoding.inputIds[1])
+        assertEquals(2088L, encoding.inputIds[2])
+        assertEquals(102L, encoding.inputIds[3])
+        assertEquals(0L, encoding.inputIds[4])  // [PAD]
         // attention mask
-        assertEquals(1, encoding.attentionMask[0])
-        assertEquals(1, encoding.attentionMask[1])
-        assertEquals(1, encoding.attentionMask[2])
-        assertEquals(1, encoding.attentionMask[3])
-        assertEquals(0, encoding.attentionMask[4])
+        assertEquals(1L, encoding.attentionMask[0])
+        assertEquals(1L, encoding.attentionMask[1])
+        assertEquals(1L, encoding.attentionMask[2])
+        assertEquals(1L, encoding.attentionMask[3])
+        assertEquals(0L, encoding.attentionMask[4])
     }
 
     @Test
     fun `encode empty string returns cls sep`() {
         val encoding = tokenizer.encode("")
-        assertEquals(101, encoding.inputIds[0])
-        assertEquals(102, encoding.inputIds[1])
+        assertEquals(101L, encoding.inputIds[0])
+        assertEquals(102L, encoding.inputIds[1])
     }
 
     @Test
     fun `wordpiece subword tokenization with hash hash`() {
         val encoding = tokenizer.encode("testing")
-        // test + ##ing
-        assertEquals(101, encoding.inputIds[0])
-        // test should be [UNK]=100, ##ing=3000
-        // Actually the minimal vocab doesn't have "test", but has "##ing"
-        // So "testing" -> unk + ##ing ?
-        // wordPiece splits: t e s t i n g, start=0 end=7
-        // tries testing, testin, testi, test -> not found
-        // tests... eventually end=0, shouldn't happen
-        // Actually with wordPiece algorithm, it tries the longest prefix
-        // For "testing", the loop goes:
-        // start=0, end=7, substr="testing" not in vocab
-        // end=6, "testin" not in vocab
-        // end=5, "testi" not in vocab
-        // end=4, "test" not in vocab (only has "##ing")
-        // end=3, "tes" not in vocab
-        // end=2, "te" not in vocab
-        // end=1, "t" not in vocab
-        // No match found -> return [UNK]
-        // So "testing" -> [CLS][UNK][SEP]
-        assertEquals(100, encoding.inputIds[1])
+        assertEquals(101L, encoding.inputIds[0])
+        assertEquals(100L, encoding.inputIds[1])
     }
 
     @Test
     fun `chinese characters are not split`() {
         val encoding = tokenizer.encode("\u4f60\u597d")
-        assertEquals(101, encoding.inputIds[0])
-        assertEquals(2000, encoding.inputIds[1])
-        assertEquals(102, encoding.inputIds[2])
+        assertEquals(101L, encoding.inputIds[0])
+        assertEquals(2000L, encoding.inputIds[1])
+        assertEquals(102L, encoding.inputIds[2])
     }
 
     @Test
@@ -73,9 +55,9 @@ class BertTokenizerTest {
         val encoding = tokenizer.encode(longText)
         assertEquals(512, encoding.inputIds.size)
         // 512th token should be [SEP]
-        assertEquals(102, encoding.inputIds[511])
+        assertEquals(102L, encoding.inputIds[511])
         // first token is [CLS]
-        assertEquals(101, encoding.inputIds[0])
+        assertEquals(101L, encoding.inputIds[0])
     }
 
     @Test
@@ -91,7 +73,7 @@ class BertTokenizerTest {
         val idx = encoding.inputIds.indexOfFirst { it == 0L }
         if (idx >= 0) {
             for (i in idx until encoding.attentionMask.size) {
-                assertEquals(0, encoding.attentionMask[i].toInt())
+                assertEquals(0L, encoding.attentionMask[i])
             }
         }
     }
